Separated Bullet from towers.rs
